+++ Assets/Hyo/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 993b919c4831bb14db6548dd52b4f276
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
