+++ Assets/Hyo/Materials/Black.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e4420a6fb8c6170498c85c2f05287650
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
