+++ Assets/Hyo/Materials/Blue.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ab5447f3ea3a3884d892ca045ae3e59f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
