+++ Assets/Hyo/Materials/Green.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4d40f071cf68ba5488de821c4b05e19c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
