+++ Assets/Hyo/Materials/Red.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 06d6c23bbfc18bf40a4b2370f4397387
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
