+++ Assets/Hyo/Materials/Yellow.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b8c4f360507da90479140508983992a6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
