+++ Assets/Hyo/Prefabs/ItemKey.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 45327abb88122754aa05f5313fff56b1
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
