+++ Assets/Hyo/Prefabs/ItemRepellent.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 93a4c2e0ff9f80d42886c62a07a24247
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
