+++ Assets/Hyo/Prefabs/ItemSpeedPotion.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 93e22f65e9be26a4a9dffd7181ecc527
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
