+++ Assets/Hyo/Script.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e76c8edc0118f7c489c1bbe1c3f50d3c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
