+++ Assets/Hyo/Script/ItemObjectMovement.cs
@@ -1,0 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class ItemObjectMovement : MonoBehaviour
+{
+    private Vector3 startPos;
+
+    void Start()
+    {
+        startPos = transform.position;
+    }
+
+    private void Update()
+    {
+        transform.Rotate(Vector3.up * 25 * Time.deltaTime);
+
+        float newY = Mathf.Sin(1 * Time.time) * 0.2f;
+        transform.position = startPos + new Vector3(0, newY, 0);
+    }
+}
