+++ Assets/Hyo/Script/ItemObjectMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4d620abc4bd805143ac0dbe320c62848
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
