+++ Assets/Hyo/Script/ItemPickup.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class ItemPickup : MonoBehaviour
+{
+    public Item.ItemType type;
+    public int value = 1;
+}
