+++ Assets/Hyo/Script/ItemPickup.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b252eae1730a7194b9b42d50540c6e5d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
